code was optimized"

+++ src/main/java/core/basesyntax/AreaCounter.java
@@ -1,5 +1,5 @@
 package core.basesyntax;
 
 public interface AreaCounter {
-    float countArea();
+    double countArea();
 }
